Don't create an empty <summary> in Atom feeds
When creating an AtomEntry from an Item, only set Summary if
item.Description is non-empty, like we already do for Content.

Currently we always emit a <summary> element, even if the item has a
non-empty <content> element. Having an empty <summary> in this case may
be confusing for feed consumers.

The Atom RFC explicitly says that a <summary> is not required in general.
(There are a couple special cases where it is required, but they aren't
relevant here.)

https://tools.ietf.org/html/rfc4287#section-4.1.1.1

Fixes #82